Move `WorkerMap` into its own module
